# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Gram Pragati Setu API
  version: 1.0.0
  description: Backend API for Gram Pragati Setu web portal
servers:
  - url: http://localhost:5000/api
    description: Development server
paths:
...
  /projects:
    get:
      summary: List projects
      tags: [Projects]
      security:
        - bearerAuth: []
      parameters:
        - name: village_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of projects
  /sync/reports:
    post:
      summary: Sync reports from mobile app
      tags: [Sync]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reports, volunteer_id]
              properties:
                volunteer_id:
                  type: string
                reports:
                  type: array
                  items:
                    type: object
                    properties:
                      client_id:
                        type: string
                      checkpoint_id:
                        type: integer
                      submitted_at:
                        type: string
                        format: date-time
                      media_items:
                        type: array
      responses:
        '200':
          description: Sync successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  synced_map:
                    type: object
                  synced_count:
                    type: integer
  /sync/media/presign:
    post:
      summary: Get presigned URL for media upload
      tags: [Sync]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [file_name, file_type]
              properties:
                file_name:
                  type: string
                file_type:
                  type: string
                file_size:
                  type: integer
...
  /analytics/projects:
    get:
      summary: Project-level analytics and checkpoint aggregates
      tags: [Analytics]
      security:
        - bearerAuth: []
      parameters:
        - name: project_type
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Project analytics
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT